Fix ARM64-RTOS build for aarch64-elf-gcc toolchain
- Update CROSS_COMPILE to aarch64-elf- (Homebrew toolchain)
- Add missing function declarations in rtos_types.h for spinlock and task APIs
- Fix MMIO macro pointer cast warnings with uintptr_t
- Add __attribute__((unused)) for intentionally unused parameters/variables
- Fix strict-aliasing violations in OPC UA float encoding using memcpy
- Export socket_table and socket_lock for TCP module linkage
- Remove duplicate timer driver (functionality in kernel/interrupt.c)
- Fix linker flags (remove gcc-only options from ld)
- Add missing CONFIG_UART_COUNT and CONFIG_UART_CLOCK defines

+++ ARM64-RTOS/kernel/interrupt.c
@@ -47,8 +47,8 @@
 #define GICC_HPPIR          (CONFIG_GICC_BASE + 0x018)
 
 /* Memory-Mapped I/O */
-#define mmio_read32(addr)       (*(volatile uint32_t *)(addr))
-#define mmio_write32(addr, val) (*(volatile uint32_t *)(addr) = (val))
+#define mmio_read32(addr)       (*(volatile uint32_t *)(uintptr_t)(addr))
+#define mmio_write32(addr, val) (*(volatile uint32_t *)(uintptr_t)(addr) = (val))
 
 /* IRQ Handler Table */
 #define MAX_IRQS    1024
